+++ example/gen/proto/go/foo/foo.pb.gw.go
@@ -65,14 +65,14 @@ func RegisterHelloServiceHandlerFromEndpoint(ctx context.Context, mux *runtime.S
 	defer func() {
 		if err != nil {
 			if cerr := conn.Close(); cerr != nil {
-				grpclog.Infof("Failed to close conn to %s: %v", endpoint, cerr)
+				grpclog.Printf("Failed to close conn to %s: %v", endpoint, cerr)
 			}
 			return
 		}
 		go func() {
 			<-ctx.Done()
 			if cerr := conn.Close(); cerr != nil {
-				grpclog.Infof("Failed to close conn to %s: %v", endpoint, cerr)
+				grpclog.Printf("Failed to close conn to %s: %v", endpoint, cerr)
 			}
 		}()
 	}()
@@ -86,8 +86,8 @@ func RegisterHelloServiceHandler(ctx context.Context, mux *runtime.ServeMux, con
 	return RegisterHelloServiceHandlerClient(ctx, mux, NewHelloServiceClient(conn))
 }
 
-// RegisterHelloServiceHandlerClient registers the http handlers for service HelloService
-// to "mux". The handlers forward requests to the grpc endpoint over the given implementation of "HelloServiceClient".
+// RegisterHelloServiceHandler registers the http handlers for service HelloService to "mux".
+// The handlers forward requests to the grpc endpoint over the given implementation of "HelloServiceClient".
 // Note: the gRPC framework executes interceptors within the gRPC handler. If the passed in "HelloServiceClient"
 // doesn't go through the normal gRPC flow (creating a gRPC client etc.) then it will be up to the passed in
 // "HelloServiceClient" to call the correct interceptors.